+++ src/main/java/production/system/ProcesaEntrada.java
@@ -1,4 +1,4 @@
-package ProductionSystem.utils;
+package production.system;
 
 import java.text.Normalizer;
 import java.util.HashSet;
